refactor: ziofy Cache (DEV-1824) (#2512)
