+++ webapi/src/main/scala/org/knora/webapi/core/AppRouter.scala
@@ -26,6 +26,8 @@ import org.knora.webapi.responders.v2.ontology.CardinalityHandler
 import org.knora.webapi.responders.v2.ontology.OntologyHelpers
 import org.knora.webapi.settings._
 import org.knora.webapi.slice.ontology.domain.service.CardinalityService
+import org.knora.webapi.slice.ontology.domain.service.OntologyRepo
+import org.knora.webapi.slice.ontology.repo.service.OntologyCache
 
 @accessible
 trait AppRouter {
@@ -41,7 +43,9 @@ object AppRouter {
       with CardinalityHandler
       with CardinalityService
       with MessageRelay
+      with OntologyCache
       with OntologyHelpers
+      with OntologyRepo
       with PermissionUtilADM
       with ResourceUtilV2
       with StandoffTagUtilV2
@@ -58,8 +62,10 @@ object AppRouter {
           ZIO.runtime[
             CardinalityHandler
               with CardinalityService
-              with PermissionUtilADM
+              with OntologyCache
               with OntologyHelpers
+              with OntologyRepo
+              with PermissionUtilADM
               with ResourceUtilV2
               with StandoffTagUtilV2
               with ValueUtilV1
