+++ webapi/src/main/scala/org/knora/webapi/core/actors/RoutingActor.scala
@@ -30,6 +30,8 @@ import org.knora.webapi.responders.v2._
 import org.knora.webapi.responders.v2.ontology.CardinalityHandler
 import org.knora.webapi.responders.v2.ontology.OntologyHelpers
 import org.knora.webapi.slice.ontology.domain.service.CardinalityService
+import org.knora.webapi.slice.ontology.domain.service.OntologyRepo
+import org.knora.webapi.slice.ontology.repo.service.OntologyCache
 import org.knora.webapi.util.ActorUtil
 
 final case class RoutingActor(
@@ -38,8 +40,10 @@ final case class RoutingActor(
   implicit val runtime: zio.Runtime[
     CardinalityHandler
       with CardinalityService
-      with PermissionUtilADM
+      with OntologyCache
       with OntologyHelpers
+      with OntologyRepo
+      with PermissionUtilADM
       with ResourceUtilV2
       with StandoffTagUtilV2
       with ValueUtilV1
