+++ webapi/src/test/scala/org/knora/webapi/slice/ontology/repo/OntologyCacheFakeSpec.scala
@@ -18,7 +18,7 @@ object OntologyCacheFakeSpec extends ZIOSpecDefault {
   val spec: Spec[Any, Throwable] = suite("OntologyCacheFake")(
     suite("with empty cache")(test("should return empty") {
       for {
-        actual <- OntologyCache.get
+        actual <- OntologyCache.getCacheData
       } yield assertTrue(actual == OntologyCacheFake.emptyData)
     }).provide(OntologyCacheFake.emptyCache),
     suite("with empty cache when setting new data")(test("should return set cache") {
@@ -27,7 +27,7 @@ object OntologyCacheFakeSpec extends ZIOSpecDefault {
         someIri <- IriConverter.asInternalSmartIri(somePropertyIri)
         newData  = OntologyCacheFake.emptyData.copy(standoffProperties = Set(someIri))
         _       <- OntologyCacheFake.set(newData)
-        actual  <- OntologyCache.get
+        actual  <- OntologyCache.getCacheData
       } yield assertTrue(actual == newData)
     }).provide(OntologyCacheFake.emptyCache, IriConverter.layer, StringFormatter.test)
   )
